requirements.txt
gradio>=4.0.0
dashscope>=1.14.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
